tests/test_basic.py
"""
Basic tests for BSS Maitri AI Assistant
"""

import unittest
import numpy as np
from unittest.mock import patch, MagicMock
import tempfile
import json

# Mock the heavy dependencies for testing
import sys
from unittest.mock import MagicMock

# Mock transformers and torch before importing our modules
sys.modules['transformers'] = MagicMock()
sys.modules['torch'] = MagicMock()
sys.modules['accelerate'] = MagicMock()
sys.modules['bitsandbytes'] = MagicMock()

from bss_maitri.config import Config
from bss_maitri.utils import validate_audio_input, validate_video_input, format_emotion_scores

class TestConfig(unittest.TestCase):
    """Test configuration management"""
    
    def test_config_loading(self):
        """Test config file loading"""
        # Create a temporary config file
        config_data = {
            'model': {'name': 'test-model'},
            'emotion_detection': {'audio': {'sample_rate': 16000}},
            'conversation': {'max_history': 5}
        }
        
        with tempfile.NamedTemporaryFile(mode='w', suffix='.yaml', delete=False) as f:
            import yaml
            yaml.dump(config_data, f)
            config_path = f.name
        
        config = Config(config_path)
        self.assertEqual(config.model['name'], 'test-model')
        self.assertEqual(config.emotion_detection['audio']['sample_rate'], 16000)
        self.assertEqual(config.conversation['max_history'], 5)

class TestUtils(unittest.TestCase):
    """Test utility functions"""
    
    def test_validate_audio_input(self):
        """Test audio input validation"""
        # Valid audio
        valid_audio = np.random.normal(0, 0.1, 16000)  # 1 second at 16kHz
        self.assertTrue(validate_audio_input(valid_audio))
        
        # Invalid inputs
        self.assertFalse(validate_audio_input(None))
        self.assertFalse(validate_audio_input(np.array([])))
        self.assertFalse(validate_audio_input(np.zeros(100)))  # Too short
        self.assertFalse(validate_audio_input(np.zeros(16000)))  # All zeros
    
    def test_validate_video_input(self):
        """Test video input validation"""
        # Valid frame
        valid_frame = np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8)
        self.assertTrue(validate_video_input(valid_frame))
        
        # Invalid inputs
        self.assertFalse(validate_video_input(None))
        self.assertFalse(validate_video_input(np.array([])))
        self.assertFalse(validate_video_input(np.zeros((10, 10))))  # Too small
    
    def test_format_emotion_scores(self):
        """Test emotion score formatting"""
        emotions = {'joy': 0.8, 'calm': 0.2, 'stress': 0.1}
        formatted = format_emotion_scores(emotions, top_k=2)
        self.assertIn('joy: 0.80', formatted)
        self.assertIn('calm: 0.20', formatted)
        self.assertNotIn('stress', formatted)

class TestEmotionDetection(unittest.TestCase):
    """Test emotion detection components"""
    
    def setUp(self):
        """Setup test fixtures"""
        self.sample_emotions = {
            'calm': 0.3, 'stress': 0.4, 'joy': 0.2, 'sadness': 0.1
        }
    
    def test_emotion_normalization(self):
        """Test emotion score normalization"""
        # This would test the actual EmotionDetector._normalize_emotions method
        # For now, just test the concept
        emotions = {'joy': 0.8, 'calm': 0.4}
        total = sum(emotions.values())
        normalized = {k: v/total for k, v in emotions.items()}
        
        self.assertAlmostEqual(sum(normalized.values()), 1.0, places=5)
    
    def test_dominant_emotion_detection(self):
        """Test dominant emotion identification"""
        dominant_emotion = max(self.sample_emotions.items(), key=lambda x: x[1])
        self.assertEqual(dominant_emotion[0], 'stress')
        self.assertEqual(dominant_emotion[1], 0.4)

class TestConversationEngine(unittest.TestCase):
    """Test conversation engine components"""
    
    def test_intervention_templates(self):
        """Test intervention template structure"""
        # Test that intervention templates have required keys
        template_structure = {
            'greeting': "Test greeting",
            'techniques': ["Technique 1", "Technique 2"],
            'follow_up': "Test follow-up"
        }
        
        required_keys = ['greeting', 'techniques', 'follow_up']
        for key in required_keys:
            self.assertIn(key, template_structure)
    
    def test_response_strategy_determination(self):
        """Test response strategy logic"""
        # Test keywords that should trigger supportive responses
        support_keywords = ['stress', 'worried', 'anxious', 'tired', 'help']
        
        test_inputs = [
            "I'm feeling stressed",
            "I'm worried about the mission",
            "Can you help me?"
        ]
        
        for test_input in test_inputs:
            has_support_keyword = any(keyword in test_input.lower() for keyword in support_keywords)
            self.assertTrue(has_support_keyword)

@patch('bss_maitri.models.gemma_model.AutoTokenizer')
@patch('bss_maitri.models.gemma_model.AutoModelForCausalLM')
@patch('bss_maitri.models.gemma_model.pipeline')
class TestIntegration(unittest.TestCase):
    """Integration tests with mocked dependencies"""
    
    def test_basic_assistant_initialization(self, mock_pipeline, mock_model, mock_tokenizer):
        """Test basic assistant initialization with mocked dependencies"""
        
        # Mock the heavy components
        mock_tokenizer.from_pretrained.return_value = MagicMock()
        mock_model.from_pretrained.return_value = MagicMock()
        mock_pipeline.return_value = MagicMock()
        
        # This test would verify that the assistant can be initialized
        # without actually loading the heavy models
        pass  # Actual test would require more complex mocking

if __name__ == '__main__':
    # Run the tests
    unittest.main(verbosity=2)