requirements.txt
# Core ML and AI dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
speech-recognition>=3.10.0
pyaudio>=0.2.11

# Computer vision and video processing
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Model serving and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.16.0

# Configuration
pyyaml>=6.0.1
hydra-core>=1.3.0